Improve docs and convention
