+++ docs/pages/docs/routing/index.mdx
@@ -3,13 +3,13 @@ import {Card} from 'nextra-theme-docs';
 
 # Internationalized routing
 
-With the introduction of the App Router, Next.js no longer provides integrated i18n routing. To fill in the gap, `next-intl` provides the two necessary pieces:
+When you provide content in multiple languages, you want to make the content available under distinct URLs (e.g. `ca.example.com/en/about`). `next-intl` provides the building blocks to set up internatinoalized routing as well as the navigation APIs to enable you to link between pages.
 
 <div className="mt-8 flex flex-col gap-4 md:w-1/2">
   <Card
     arrow
     icon={<BoltIcon />}
-    title="Internationalized routing middleware"
+    title="Routing middleware"
     href="/docs/routing/middleware"
   />
   <Card
@@ -19,3 +19,5 @@ With the introduction of the App Router, Next.js no longer provides integrated i
     href="/docs/routing/navigation"
   />
 </div>
+
+Note that these features are only relevant if you use the App Router. If you're using [`next-intl` with the Pages Router](/docs/getting-started/pages-router), you can use the [built-in capabilities from Next.js](https://nextjs.org/docs/pages/building-your-application/routing/internationalization).
