+++ docs/pages/index.mdx
@@ -24,7 +24,7 @@ import Hero from 'components/Hero';
 <PartnerBanner intro="Presented by" />
 
 <Section
-  title="Minimalistic, yet complete"
+  title="Batteries included, minimalistic API"
   description="Internationalization (i18n) is an essential part of the user experience, therefore next‑intl gives you everything you need to get language nuances right in your Next.js apps."
 >
   <div className="grid grid-cols-1 lg:grid-cols-2 gap-6 xl:gap-20">
