+++ examples/example-next-13/src/middleware.ts
@@ -1,5 +1,5 @@
 import createMiddleware from 'next-intl/middleware';
-import {pathnames, locales} from './i18n';
+import {pathnames, locales} from './navigation';
 
 export default createMiddleware({
   defaultLocale: 'en',
